+++ src/Sidebar/Katalog/KatalogNavigering/Navigeringsliste/NavigeringslisteFunksjoner/getKey.js
@@ -1,0 +1,13 @@
+import språk from "Funksjoner/språk";
+const getKey = (node, sorterPåKode) => {
+  if (node.sortering) return node.sortering;
+  if (sorterPåKode)
+    return node.kode
+      .replace("+", "Z")
+      .split(/-/)
+      .map(e => e.padStart(5, "0"));
+
+  return språk(node.tittel);
+};
+
+export default { getKey };
